Ken's comments
