+++ openmdao/components/tests/test_meta_model_structured_comp.py
@@ -2,6 +2,7 @@
 Unit tests for the structured metamodel component.
 """
 import unittest
+import inspect
 
 import numpy as np
 from numpy.testing import assert_almost_equal
@@ -1085,6 +1086,32 @@ class TestMetaModelStructuredPython(unittest.TestCase):
         prob.run_model()
 
         self.run_and_check_derivs(prob)
+
+    def test_snopt_analysis_error(self):
+
+      x_tr = np.linspace(0, 2*np.pi, 100)
+      y_tr = np.sin(x_tr)
+
+      p = om.Problem(model=om.Group())
+
+      p.driver = om.pyOptSparseDriver(optimizer='SNOPT')
+
+      mm = om.MetaModelStructuredComp(extrapolate=False)
+      mm.add_input('x', val=1.0, training_data=x_tr)
+      mm.add_output('y', val=1.0, training_data=y_tr)
+      p.model.add_subsystem('interp', mm, promotes_inputs=['x'], promotes_outputs=['y'])
+
+      p.model.add_objective('y', scaler=-1)
+      p.model.add_design_var('x', lower=6, upper=10)
+
+      p.set_solver_print(level=0)
+      p.setup()
+
+      p.set_val('x', 0.75)
+
+      msg = "Analysis Error: Line 205 of file {}".format(inspect.getsourcefile(om.MetaModelStructuredComp))
+      with assert_warning(UserWarning, msg):
+          p.run_driver()
 
 
 @unittest.skipIf(not scipy_gte_019, "only run if scipy>=0.19.")
